% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicator.R
\name{dedup.function}
\alias{dedup.function}
\title{Deduplicates repeat specimens from a patient sent within specified time}
\usage{
dedup.function(df, x)
}
\arguments{
\item{df}{dataframe}

\item{x}{numeric variable}
}
\description{
Deduplicates repeat specimens from a patient sent within specified time
}
